requirements.txt
streamlit
pandas
numpy
scikit-learn
Pillow
fpdf
requests
python-dotenv
